pyproject.toml
[tool.poetry]
name = "main_app"
version = "0.1.0"
description = ""
authors = ["vincentsmid <vincent.smid@greenpanda.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["jinja2"], version = "^0.109.0" }
piccolo = { extras = ["sqlite"], version = "^1.3.0" }
passlib = "^1.7.4"
bcrypt = "^4.1.2"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.3"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
httpx = "^0.26.0"
jinja2 = "^3.1.3"
aiofiles = "^23.2.1"
fastapi-cors = "^0.0.6"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
piccolo-admin = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pre-commit = "^3.6.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
